fix: update import paths after moving ClassicBrowserService to services/browser
- Fixed all imports referencing ClassicBrowserService to use new path
- Updated relative imports within ClassicBrowserService itself
- Updated 26 files across the codebase with correct import paths
