+++ electron/ipc/classicBrowserSetBounds.ts
@@ -1,6 +1,6 @@
 import { ipcMain, IpcMainEvent } from 'electron';
 import { CLASSIC_BROWSER_SET_BOUNDS } from '../../shared/ipcChannels';
-import { ClassicBrowserService } from '../../services/ClassicBrowserService';
+import { ClassicBrowserService } from '../../services/browser/ClassicBrowserService';
 
 const logger = {
   debug: (...args: any[]) => console.log('[IPCClassicBrowserSetBounds]', ...args),
