+++ electron/ipc/syncWindowStackOrder.ts
@@ -1,5 +1,5 @@
 import { ipcMain } from 'electron';
-import { ClassicBrowserService } from '../../services/ClassicBrowserService';
+import { ClassicBrowserService } from '../../services/browser/ClassicBrowserService';
 import { logger } from '../../utils/logger';
 import { SYNC_WINDOW_STACK_ORDER } from '../../shared/ipcChannels';
