+++ services/NotebookCompositionService.ts
@@ -4,7 +4,7 @@ import * as path from 'path';
 import { v4 as uuidv4 } from 'uuid';
 import { NotebookService } from './NotebookService';
 import { ObjectModel } from '../models/ObjectModel';
-import { ClassicBrowserService } from './ClassicBrowserService';
+import { ClassicBrowserService } from './browser/ClassicBrowserService';
 import { WindowMeta, ClassicBrowserPayload, TabState } from '../shared/types';
 import { logger } from '../utils/logger';
 import { BaseService } from './base/BaseService';
